Make FS slashes more consistent and *nix style. Abandon the fixed idea that a long can represent a pointer. Thank you, Bill. Add a new intptr_t type since long no longer works for this. Stop obsessive last error clearing. Add RU_WALK_UNIX_SLASHES because the callbacks get native paths. Add Mutex debugging instrumentation. Start accounting for the fact that Windows mucks with line endings during I/O. Settle on binary mode where implied. Fix folder walk bugs. Fix timestamp bugs. Guard against accidental recursive use of ruMakeLogMsgV.

+++ include/regify-util.h
@@ -328,12 +328,12 @@ typedef int64_t usec_t;
 /**
  * \brief A pointer sized integer type for collections like \ref ruMap or \ref ruList.
  */
-typedef long ru_int;
+typedef intptr_t ru_int;
 
 /**
  * \brief A pointer sized unsigned integer type for collections like \ref ruMap or \ref ruList.
  */
-typedef unsigned long ru_uint;
+typedef uintptr_t ru_uint;
 
 /**
  * \brief Abstracted version of size_t.
